CUMULUS-2801 -- Remove Dynamo references from related code, update tests (#2958)
