+++ packages/api/tests/models/granules/test-granules-model.js
@@ -4,7 +4,6 @@ const test = require('ava');
 const sinon = require('sinon');
 
 const awsServices = require('@cumulus/aws-client/services');
-const s3Utils = require('@cumulus/aws-client/S3');
 const StepFunctions = require('@cumulus/aws-client/StepFunctions');
 const launchpad = require('@cumulus/launchpad-auth');
 const { randomString } = require('@cumulus/common/test-utils');
@@ -12,7 +11,7 @@ const { CMR } = require('@cumulus/cmr-client');
 const { DefaultProvider } = require('@cumulus/common/key-pair-provider');
 
 const Granule = require('../../../models/granules');
-const { fakeFileFactory, fakeGranuleFactoryV2 } = require('../../../lib/testUtils');
+const { fakeGranuleFactoryV2 } = require('../../../lib/testUtils');
 
 let fakeExecution;
 let testCumulusMessage;
@@ -90,196 +89,6 @@ test.after.always(async () => {
   }).promise();
   await new Granule().deleteTable();
   sandbox.restore();
-});
-
-test('files existing at location returns empty array if no files exist', async (t) => {
-  const filenames = [
-    'granule-file-1.hdf',
-    'granule-file-2.hdf',
-  ];
-
-  const sourceBucket = 'test-bucket';
-  const destBucket = 'dest-bucket';
-
-  const sourceFiles = filenames.map(
-    (name) =>
-      fakeFileFactory({
-        name,
-        bucket: sourceBucket,
-        key: `origin/${name}`,
-      })
-  );
-
-  const destinationFilepath = 'destination';
-
-  const destinations = [
-    {
-      regex: '.*.hdf$',
-      bucket: destBucket,
-      key: destinationFilepath,
-    },
-  ];
-
-  const granule = {
-    files: sourceFiles,
-  };
-
-  const granulesModel = new Granule();
-
-  const filesExisting = await granulesModel.getFilesExistingAtLocation(granule, destinations);
-
-  t.deepEqual(filesExisting, []);
-});
-
-test('files existing at location returns both files if both exist', async (t) => {
-  const filenames = [
-    'granule-file-1.hdf',
-    'granule-file-2.hdf',
-  ];
-
-  const sourceBucket = 'test-bucket';
-  const destBucket = randomString();
-
-  await awsServices.s3().createBucket({ Bucket: destBucket });
-
-  const sourceFiles = filenames.map(
-    (fileName) => fakeFileFactory({ fileName, bucket: sourceBucket })
-  );
-
-  const destinations = [
-    {
-      regex: '.*.hdf$',
-      bucket: destBucket,
-    },
-  ];
-
-  const dataSetupPromises = filenames.map(async (filename) => {
-    const params = {
-      Bucket: destBucket,
-      Key: filename,
-      Body: 'test',
-    };
-    return await awsServices.s3().putObject(params);
-  });
-
-  await Promise.all(dataSetupPromises);
-
-  const granule = {
-    files: sourceFiles,
-  };
-
-  const granulesModel = new Granule();
-
-  const filesExisting = await granulesModel.getFilesExistingAtLocation(granule, destinations);
-
-  t.deepEqual(filesExisting, sourceFiles);
-
-  await s3Utils.recursivelyDeleteS3Bucket(destBucket);
-});
-
-test('files existing at location returns only file that exists', async (t) => {
-  const filenames = [
-    'granule-file-1.hdf',
-    'granule-file-2.hdf',
-  ];
-
-  const sourceBucket = 'test-bucket';
-  const destBucket = randomString();
-
-  await awsServices.s3().createBucket({ Bucket: destBucket });
-
-  const sourceFiles = filenames.map(
-    (fileName) => fakeFileFactory({ fileName, bucket: sourceBucket })
-  );
-
-  const destinations = [
-    {
-      regex: '.*.hdf$',
-      bucket: destBucket,
-      filepath: '',
-    },
-  ];
-
-  const params = {
-    Bucket: destBucket,
-    Key: filenames[1],
-    Body: 'test',
-  };
-  await awsServices.s3().putObject(params);
-
-  const granule = {
-    files: sourceFiles,
-  };
-
-  const granulesModel = new Granule();
-
-  const filesExisting = await granulesModel.getFilesExistingAtLocation(granule, destinations);
-
-  t.deepEqual(filesExisting, [sourceFiles[1]]);
-
-  await s3Utils.recursivelyDeleteS3Bucket(destBucket);
-});
-
-test('files existing at location returns only file that exists with multiple destinations', async (t) => {
-  const filenames = [
-    'granule-file-1.txt',
-    'granule-file-2.hdf',
-  ];
-
-  const sourceBucket = 'test-bucket';
-  const destBucket1 = randomString();
-  const destBucket2 = randomString();
-
-  await Promise.all([
-    awsServices.s3().createBucket({ Bucket: destBucket1 }),
-    awsServices.s3().createBucket({ Bucket: destBucket2 }),
-  ]);
-
-  const sourceFiles = filenames.map(
-    (fileName) => fakeFileFactory({ fileName, bucket: sourceBucket })
-  );
-
-  const destinations = [
-    {
-      regex: '.*.txt$',
-      bucket: destBucket1,
-      filepath: '',
-    },
-    {
-      regex: '.*.hdf$',
-      bucket: destBucket2,
-      filepath: '',
-    },
-  ];
-
-  let params = {
-    Bucket: destBucket1,
-    Key: filenames[0],
-    Body: 'test',
-  };
-  await awsServices.s3().putObject(params);
-
-  params = {
-    Bucket: destBucket2,
-    Key: filenames[1],
-    Body: 'test',
-  };
-  await awsServices.s3().putObject(params);
-
-  const granule = {
-    files: sourceFiles,
-  };
-
-  const granulesModel = new Granule();
-
-  const filesExisting = await granulesModel.getFilesExistingAtLocation(granule, destinations);
-
-  t.deepEqual(filesExisting, sourceFiles);
-
-  await Promise.all([
-    s3Utils.recursivelyDeleteS3Bucket(destBucket1),
-    s3Utils.recursivelyDeleteS3Bucket(destBucket2),
-  ]);
 });
 
 test('get() will translate an old-style granule file into the new schema', async (t) => {
